campus enum class

+++ src/transactionmanager/Checking.java
@@ -2,7 +2,8 @@ package transactionmanager;
 
 public class Checking extends Account {
     public static final double ANNUALINTERESTRATE = 0.01;
-    public static final int MONTHLYFEE = 12;
+    public static final double MONTHLYFEE = 12;
+
