+++ src/transactionmanager/CollegeChecking.java
@@ -3,5 +3,5 @@ package transactionmanager;
 public class CollegeChecking extends Checking {
     private Campus campus; //campus code
     public static final double ANNUALINTERESTRATE = 0.01;
-    public static final int MONTHLYFEE = 0;
+    public static final double MONTHLYFEE = 0;
 }
