+++ src/transactionmanager/MoneyMarket.java
@@ -3,5 +3,5 @@ package transactionmanager;
 public class MoneyMarket extends Savings{
     private int withdrawal; //number of withdrawals
     public static final double ANNUALINTERESTRATE = 0.045;
-    public static final int MONTHLYFEE = 25;
+    public static final double MONTHLYFEE = 25;
 }
