+++ src/transactionmanager/Savings.java
@@ -3,5 +3,5 @@ package transactionmanager;
 public class Savings extends Account{
     protected boolean isLoyal; //loyal customer status
     public static final double ANNUALINTERESTRATE = 0.04;
-    public static final int MONTHLYFEE = 25;
+    public static final double MONTHLYFEE = 25;
 }
